Change AverageRank property type and fix its bug
- Bug of setting value to AverageRank fixed
- AverageRank property of Professor model changed to float,
  dtos also updated and migrations created and database updated

+++ src/KodoomOstad.Entities/Models/Professor.cs
@@ -12,7 +12,7 @@ namespace KodoomOstad.Entities.Models
 
         public ProfessorRank Rank { get; set; }
 
-        public int AverageRate { get; set; }
+        public float AverageRate { get; set; }
 
         public int FacultyId { get; set; }
         public Faculty Faculty { get; set; }
