+++ src/KodoomOstad.WebApi/Models/Professors/ProfessorOutputDto.cs
@@ -11,7 +11,7 @@ namespace KodoomOstad.WebApi.Models.Professors
 
         public string Rank { get; set; }
 
-        public int AverageRate { get; set; }
+        public float AverageRate { get; set; }
 
         public string Avatar { get; set; }
